Update FindCorrelation_2.R
**Pairwise correlation between variables**

We first try to find pairwise correlation between all the variables in our data frame. But we remove our target phenotype "DMI" and other binary variables before actually running pairwise correlation. The objective of doing this is to find potential confounders which should be accounted for in the differential gene expression analysis. However highest correlating variables can be used to predict a phenotype accurately.